# services/mapi-xdg/docs/indexer/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dogecoin - Blockchain Indexer API",
    "description": "Dogecoin indexer endpoints: addresses, dunes, inscriptions and DRC-20. The same binary also proxies a set of node RPC endpoints (blocks, mempool, chain info) documented in the repository README.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.txt"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local instance"
    }
  ],
  "paths": {
...
    "/assets/dunes/{dune}": {
      "get": {
        "tags": [
          "Dunes"
        ],
        "summary": "Dunes Info",
        "description": "Returns information about the specified Dune, including etching information,\ncurrent supply and number of holders.",
        "operationId": "info_by_dune",
        "parameters": [
          {
            "name": "dune",
            "in": "path",
            "description": "Dune, specified either by the Dune ID (etching block number and transaction index) or name (spaced or un-spaced)",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2519999:31"
          }
        ],
        "responses": {
          "200": {
            "description": "Requested data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimestampedDuneInfo"
                },
                "example": {
                  "data": {
                    "divisibility": 8,
                    "etching_cenotaph": false,
                    "etching_height": 5280961,
                    "etching_tx": "39a7cc8155a9e68692ee280e2dfb5a4007cfa5f83f2801abf10fb2682ee3fabc",
                    "id": "5280961:140",
                    "max_supply": "1000000000000.00000000",
                    "mints": 2,
                    "name": "TESTTESTTEEST",
                    "spaced_name": "TEST•TEST•TEEST",
                    "symbol": "a",
                    "terms": {
                      "amount_per_mint": "1000.00000000",
                      "end_height": "5290000",
                      "end_offset": null,
                      "mint_txs_cap": "1000000000",
                      "start_height": "5280850",
                      "start_offset": null
                    },
                    "unique_holders": 2
                  },
                  "last_updated": {
                    "block_hash": "304f07cefebe68472f2ceb02706f9a3b5a9e9ecb3534ec34923799c8ecd7f27d",
                    "block_height": 5303197
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameters"
          },
          "404": {
            "description": "Requested entity not found on-chain"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/assets/dunes/{dune}/holders": {
      "get": {
        "tags": [
          "Dunes"
        ],
        "summary": "Holders by Dune",
        "description": "List of all addresses that hold the specified Dune, with the respective amounts.",
        "operationId": "holders_by_dune",
        "parameters": [
          {
            "name": "dune",
            "in": "path",
            "description": "Dune, specified either by the Dune ID (etching block number and transaction index) or name (spaced or un-spaced)",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "5095478:8"
          },
          {
            "name": "count",
            "in": "query",
            "description": "The max number of results per page",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "type": "integer",
                  "default": 100,
                  "minimum": 0
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Pagination cursor string, use the cursor included in a page of results to fetch the next page",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Requested data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedDuneHolder"
                },
                "example": {
                  "data": [
                    {
                      "address": "D8UC42uehZdBYAfMBJumQKuBNs9vvarxtt",
                      "balance": "9000000.000000000000000000",
                      "script_pubkey": "76a9142cbe0e5f8f21b1b1a5d1d0d9e3a2b4c5d6e7f80988ac"
                    },
                    {
                      "address": "DEvbE2KyPKAzDsj511wijyF1NSBN9qMcix",
                      "balance": "420000.000000000000000000",
                      "script_pubkey": "76a914e52f1c1b6a2f7d3c4b5a69788736251409f3ab8788ac"
                    }
                  ],
                  "last_updated": {
                    "block_hash": "1c3e7cd9e46bd6d0adb1ae0b52ec1a1ddfaa0cb61a41a1b1c26f0d1b0f42a06c",
                    "block_height": 5150534
                  },
                  "next_cursor": "19FwuaejD9hE1R4ckTQKaqe0ecA"
                }
              }
            }
          },
          "400": {
            "description": "Malformed query parameters"
          },
          "404": {
            "description": "Requested entity not found on-chain"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DuneHolder": {
        "type": "object",
        "required": [
          "script_pubkey",
          "balance"
        ],
        "properties": {
          "address": {
            "type": "string",
            "nullable": true
          },
          "balance": {
            "type": "string"
          },
          "script_pubkey": {
            "type": "string"
          }
        }
      },
...
      "DuneInfo": {
        "type": "object",
        "required": [
          "id",
          "etching_cenotaph",
          "etching_tx",
          "etching_height",
          "name",
          "spaced_name",
          "divisibility",
          "terms",
          "circulating_supply",
          "mints",
          "unique_holders"
        ],
        "properties": {
          "circulating_supply": {
            "type": "string"
          },
          "divisibility": {
            "type": "integer",
            "format": "int32",
            "description": "If no divisibility was specified, then this equals 0",
            "minimum": 0
          },
          "etching_cenotaph": {
            "type": "boolean"
          },
          "etching_height": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "etching_tx": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "max_supply": {
            "type": "string",
            "nullable": true
          },
          "mints": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "name": {
            "type": "string"
          },
          "spaced_name": {
            "type": "string"
          },
          "symbol": {
            "type": "string",
            "nullable": true
          },
          "terms": {
            "$ref": "#/components/schemas/Terms"
          },
          "unique_holders": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "LastUpdated": {
        "type": "object",
        "required": [
          "block_hash",
          "block_height"
        ],
        "properties": {
          "block_hash": {
            "type": "string"
          },
          "block_height": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "PaginatedDuneHolder": {
        "type": "object",
        "required": [
          "data",
          "last_updated"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DuneHolder"
            }
          },
          "last_updated": {
            "$ref": "#/components/schemas/LastUpdated"
          },
          "next_cursor": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "Terms": {
        "type": "object",
        "properties": {
          "amount_per_mint": {
            "type": "string",
            "nullable": true
          },
          "end_height": {
            "type": "string",
            "nullable": true
          },
          "end_offset": {
            "type": "string",
            "nullable": true
          },
          "mint_txs_cap": {
            "type": "string",
            "nullable": true
          },
          "start_height": {
            "type": "string",
            "nullable": true
          },
          "start_offset": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "TimestampedDuneInfo": {
        "type": "object",
        "required": [
          "data",
          "last_updated"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DuneInfo"
          },
          "last_updated": {
            "$ref": "#/components/schemas/LastUpdated"
          }
        }
      },
...
    }
  }
}